Finished app

+++ app/Http/Controllers/ListingController.php
@@ -46,6 +46,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -59,6 +61,13 @@ class ListingController extends Controller
     // Update listing
     public function update(Request $request, Listing $listing)
     {
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id())
+        {
+            abort(403, 'Unauthorized action!');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -83,7 +92,18 @@ class ListingController extends Controller
     // Delete listing
     public function delete(Listing $listing)
     {
+        if($listing->user_id != auth()->id())
+        {
+            abort(403, 'Unauthorized action!');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
